+++ packages/config-utils/package.json
@@ -16,12 +16,6 @@
         "url": "https://github.com/RedHatInsights/frontend-components/issues"
     },
     "homepage": "https://github.com/RedHatInsights/frontend-components#readme",
-    "dependencies": {
-        "deepmerge": "^4.0.0",
-        "glob": "^7.0.0",
-        "js-yaml": "^4.0.0",
-        "jws": "^4.0.0"
-    },
     "peerDependencies": {
         "webpack": "^5.0.0"
     }
