+++ packages/config/index.js
@@ -1,6 +1,6 @@
-const { sync } = require('glob');
 const config = require('./src/config');
 const plugins = require('./src/plugins');
+const { sync } = require('glob');
 
 const gitRevisionPlugin = new(require('git-revision-webpack-plugin'))({
     branch: true
