+++ packages/config/package.json
@@ -26,8 +26,12 @@
         "css-loader": "^5.2.6",
         "file-loader": "^6.2.0",
         "git-revision-webpack-plugin": "^3.0.6",
+        "glob": "^7.0.0",
+        "html-replace-webpack-plugin": "^2.6.0",
         "html-webpack-plugin": "^5.3.1",
         "mini-css-extract-plugin": "^1.6.0",
+        "js-yaml": "^4.0.0",
+        "jws": "^4.0.0",
         "path-browserify": "^1.0.1",
         "process": "^0.11.10",
         "sass": "^1.34.1",
